Распределила функции по ЕТ
