+++ shagieva.dinara/T1/dataFunctions.cpp
@@ -1,8 +1,10 @@
-#include "parseFigure.hpp"
+#include "dataFunctions.hpp"
 #include "rectangle.hpp"
 #include "complexquad.hpp"
 #include "regular.hpp"
+#include "base-types.hpp"
 #include <stdexcept>
+#include <iostream>
 
 void shagieva::parseFigure(std::string figureType, std::istream & input, shagieva::Shape ** shapes, size_t & shapeCount)
 {
@@ -55,5 +57,18 @@ void shagieva::parseFigure(std::string figureType, std::istream & input, shagiev
   else
   {
     throw std::invalid_argument("Invalid figure entered");
+  }
+}
+
+void shagieva::printFramePoints(const shagieva::Shape * const * const shapes, const size_t & shapeCount)
+{
+  for (size_t i = 0; i < shapeCount; ++i)
+  {
+    shagieva::rectangle_t frame = shapes[i]->getFrameRect();
+    std::cout << frame.pos.x - frame.width / 2 << " "
+              << frame.pos.y - frame.height / 2 << " "
+              << frame.pos.x + frame.width / 2 << " "
+              << frame.pos.y + frame.height / 2
+              << ((i == (shapeCount - 1)) ? "\n" : " ");
   }
 }
