+++ shagieva.dinara/T1/dataFunctions.hpp
@@ -1,5 +1,5 @@
-#ifndef PARSE_FIGURE_HPP
-#define PARSE_FIGURE_CPP
+#ifndef DATA_FUNCTIONS_HPP
+#define DATA_FUNCTIONS_HPP
 #include "shape.hpp"
 #include <string>
 #include <iostream>
@@ -8,6 +8,7 @@
 namespace shagieva
 {
   void parseFigure(std::string figureType, std::istream & input, Shape ** shapes, size_t & shapeCount);
-};
+  void printFramePoints(const Shape * const * const shapes, const size_t & shapeCount);
+}
 
 #endif
