--- shagieva.dinara/T1/deleteData.hpp
@@ -1,11 +1,0 @@
-#ifndef DELETE_DATA_HPP
-#define DELETE_DATA_CPP
-#include "shape.hpp"
-#include <cstddef>
-
-namespace shagieva
-{
-  void deleteData(Shape ** shapes, const size_t & shapeCount);
-};
-
-#endif
