--- shagieva.dinara/T1/printFramePoints.hpp
@@ -1,11 +1,0 @@
-#ifndef PRINT_FRAME_POINTS_HPP
-#define PRINT_FRAME_POINTS_HPP
-#include "shape.hpp"
-#include <cstddef>
-
-namespace shagieva
-{
-  void printFramePoints(const Shape * const * const shapes, const size_t & shapeCount);
-}
-
-#endif
